Update main_listener.php
Fixed "PHP Notice : Undefined index" occuring when generating personal gallery links (contact profile fields).

Explanation: If a user account has been deleted/banned but his messages still present on the forum, we will not have his account ID set in $this->user_ids. So, we first have to check if our current user is valid ($this->user_ids[$this->target] is set?), and then we check if this user have a personal gallery ($this->albums[current user] is set?).